Refactor the tests to not use a collection
