+++ src/CSnakes.Runtime.Tests/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+﻿//[assembly: CollectionBehavior(DisableTestParallelization = true)]
