--- src/CSnakes.Runtime.Tests/Converter/ConverterCollection.cs
@@ -1,6 +1,0 @@
-﻿namespace CSnakes.Runtime.Tests.Converter;
-
-[CollectionDefinition("ConversionTests")]
-public class ConverterCollection : ICollectionFixture<TestEnvironment>
-{
-}
