+++ src/CSnakes.Runtime.Tests/Converter/ConverterTestBase.cs
@@ -1,0 +1,39 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+namespace CSnakes.Runtime.Tests.Converter;
+public class ConverterTestBase : IDisposable
+{
+    protected readonly IPythonEnvironment env;
+    protected readonly IHost app;
+
+    public ConverterTestBase()
+    {
+        app = Host.CreateDefaultBuilder()
+            .ConfigureServices((context, services) =>
+            {
+                var pb = services.WithPython();
+                pb.WithHome(Environment.CurrentDirectory);
+
+                pb.FromNuGet("3.12.4").FromMacOSInstallerLocator("3.12").FromEnvironmentVariable("Python3_ROOT_DIR", "3.12.4");
+            })
+            .Build();
+
+        env = app.Services.GetRequiredService<IPythonEnvironment>();
+    }
+
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
+        GC.Collect();
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            app.Dispose();
+        }
+    }
+}
